requirements.txt
weasyprint>=61.0
markdown>=3.5
pymdown-extensions>=10.0
Pygments>=2.17
google-generativeai>=0.8.0
